+++ packages/core/styled-system/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change Log
 
+## 2.9.0
+
+### Minor Changes
+
+- [`379f347a8`](https://github.com/chakra-ui/chakra-ui/commit/379f347a891f131c39a436b5738221105300ad76)
+  Thanks [@segunadebayo](https://github.com/segunadebayo)! - Add support for
+  `aspectRatio` style prop to enable the usage of the native aspect ratio
+  property
+
+- [#7494](https://github.com/chakra-ui/chakra-ui/pull/7494)
+  [`27dcf2f56`](https://github.com/chakra-ui/chakra-ui/commit/27dcf2f5647c3323d1e6bee35db6cf1514c1b29d)
+  Thanks [@kuroppe1819](https://github.com/kuroppe1819)! - Add support for
+  nested semantic tokens in theme. It is now possible to declare semantic tokens
+  by nesting objects.
+
+  BEFORE:
+
+  ```js
+  const theme = {
+    semanticTokens: {
+      colors: {
+        "background.pressed.base": { default: "blue.800", _dark: "blue.300" },
+        "background.pressed.subtle": { default: "blue.300", _dark: "blue.700" },
+      },
+    },
+  }
+  ```
+
+  AFTER:
+
+  ```js
+  const theme = {
+    semanticTokens: {
+      colors: {
+        background: {
+          pressed: {
+            base: { default: "blue.800", _dark: "blue.300" },
+            subtle: { default: "blue.300", _dark: "blue.700" },
+          },
+        },
+      },
+    },
+  }
+  ```
+
+  This allows for cleaner grouping and organization of tokens.
+
+- [#7502](https://github.com/chakra-ui/chakra-ui/pull/7502)
+  [`49a29a238`](https://github.com/chakra-ui/chakra-ui/commit/49a29a238439242e0959d74ebf48c84411581288)
+  Thanks [@estheragbaje](https://github.com/estheragbaje)! - Add `_horizontal`
+  and `_vertical` pseudo props to style orientation data attributes
+
 ## 2.8.0
 
 ### Minor Changes